refactor(frontend): correct utility names
