+++ frontend/utils/index.ts
@@ -1,5 +1,5 @@
-import * as globals from './globals/_sanityser'
+import * as globals from './globals/_sanitizer'
 import * as images from './images/image'
-import * as handleErrors from './server/_handdleErrors'
+import * as handleErrors from './server/_handleErrors'
 
 export { globals, handleErrors, images }
