+++ frontend/utils/server/_handleErrors.ts
@@ -1,4 +1,4 @@
-//Ici, créer une fonction qui va gérer les erreurs de façon plus globale
+// Here, create a function that will handle errors in a more global way
 export function _handleError(error: unknown, message: string) {
   console.error(message, error)
   throw createError({
